requirements/requirements.txt
# Django
Django==1.9

# Configuration
django-environ==0.4.0
django-secure==1.0.1

# Python-PostgreSQL Database Adapter
psycopg2

# Forms
django-braces==1.4.0

# Search
django-haystack==2.4.1
pyelasticsearch==1.4

# REST
djangorestframework==3.3.1
django-filter==0.11.0

# Message broker
celery==3.1.19

# Taggs
django-taggit==0.17.6

# Timezones Date/time
python-dateutil==2.4.2
pytz==2015.6

# WSGI
gunicorn==19.3.0

# Other
argparse==1.2.1
django-digg-paginator==0.1.2
#wstools==0.4